requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-dotenv==1.4.2
django-rename-app==0.1.2
djangorestframework==3.12.1
djangorestframework-gis==0.16
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.1
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
